Remove DataPayload.putLong() and DataPayload.getLong() to avoid Key KEY_REMINDER_ID expected Long but value was a java.lang.Integer error from BaseBundle.
Add DataPayload.putInt() and DataPayload.getInt() back

+++ alarmscheduler/src/main/java/com/carterchen247/alarmscheduler/model/DataPayload.kt
@@ -6,12 +6,12 @@ class DataPayload {
 
     private var bundle = Bundle()
 
-    fun putLong(key: String, value: Long) {
-        bundle.putLong(key, value)
+    fun putInt(key: String, value: Int) {
+        bundle.putInt(key, value)
     }
 
-    fun getLong(key: String): Long {
-        return bundle.getLong(key)
+    fun getInt(key: String): Int {
+        return bundle.getInt(key)
     }
 
     fun putString(key: String, value: String) {
